instrutor/requirements.txt
fastmcp==2.10.2
tavily-python==0.7.12
python-dotenv>=1.0.0
pandas>=2.0.0
google-adk==1.16.0
langchain-openai==1.0.1
litellm==1.78.6
mcp==1.18.0

